third_party/blink/web_tests/wpt_internal/html/canvas/drawElementImage/privacy/iframes-ignored-onpaint.https.sub.html

<!DOCTYPE HTML>
<title>canvas.onpaint and cross-origin iframes</title>
<link rel="help" href="https://github.com/WICG/html-in-canvas">
<script src='/resources/testharness.js'></script>
<script src='/resources/testharnessreport.js'></script>
<canvas id=canvas1 width="100" height="100" layoutsubtree>
  <div id=child1>
    <iframe id=sameOrigin width=50 height=50 src="https://{{location[host]}}/wpt_internal/html/canvas/drawElementImage/privacy/support/raf-loop.html"></iframe>
  </div>
</canvas>

<canvas id=canvas2 width="100" height="100" layoutsubtree>
  <div id=child2>
    <iframe id=crossOrigin width=50 height=50 src="https://{{hosts[alt][www]}}:{{ports[h2][0]}}/wpt_internal/html/canvas/drawElementImage/privacy/support/raf-loop.html"></iframe>
  </div>
</canvas>

<script>
  let sPendingIframesToLoad = 2;

  function pendingIframeLoaded() {
    if (!--sPendingIframesToLoad) {
      promise_test(async function(t) {
        await new Promise(requestAnimationFrame);
        await new Promise(setTimeout);

        let paintCount1 = 0;
        let paintCount2 = 0;
        canvas1.onpaint = () => paintCount1++;
        canvas2.onpaint = () => paintCount2++;

        // Wait several frames to ensure the frames have time to repaint.
        await new Promise(requestAnimationFrame);
        await new Promise(requestAnimationFrame);
        await new Promise(requestAnimationFrame);
        await new Promise(r => setTimeout(r, 100));

        assert_greater_than(paintCount1, 0, "Same origin should trigger paint");
        assert_equals(paintCount2, 0, "Cross origin should not trigger paint");
      });
    }
  }

  sameOrigin.onload = pendingIframeLoaded;
  crossOrigin.onload = pendingIframeLoaded;
</script>
